Refactor : getUserFromToken 메서드 JwtUtil로 이동
